merge updates from master

+++ MSDYNV9/Xrm.Framework.CI/Xrm.Framework.CI.PowerShell.Cmdlets/PluginRegistration/PluginRegistrationHelper.cs
@@ -7,6 +7,12 @@ using System.Collections.Generic;
 using System.Linq;
 using Xrm.Framework.CI.PowerShell.Cmdlets.PluginRegistration;
 using System.IO;
+using Xrm.Framework.CI.Common.Entities;
+using System.Xml.Linq;
+using System.Xml;
+using System.Xml.XPath;
+
+using Xrm.Framework.CI.Common;
 
 namespace Xrm.Framework.CI.PowerShell.Cmdlets
 {
@@ -16,6 +22,8 @@ namespace Xrm.Framework.CI.PowerShell.Cmdlets
         private readonly PluginRepository pluginRepository;
         private readonly Action<string> logVerbose;
         private readonly Action<string> logWarning;
+        private readonly IReflectionLoader reflectionLoader;
+        private readonly IPluginRegistrationObjectFactory pluginRegistrationObjectFactory;
 
         public PluginRegistrationHelper(IOrganizationService service, CIContext xrmContext, Action<string> logVerbose, Action<string> logWarning)
         {
@@ -23,13 +31,32 @@ namespace Xrm.Framework.CI.PowerShell.Cmdlets
             this.logWarning = logWarning;
             organizationService = service;
             pluginRepository = new PluginRepository(xrmContext);
+            reflectionLoader = new ReflectionLoader();
+            pluginRegistrationObjectFactory = new PluginRegistrationObjectFactory();
         }
 
-        public Assembly GetAssemblyRegistration(string assemblyName) => pluginRepository.GetAssemblyRegistration(assemblyName);
+        public PluginRegistrationHelper(Action<string> logVerbose, Action<string> logWarning)
+        {
+            this.logVerbose = logVerbose;
+            this.logWarning = logWarning;
+            reflectionLoader = new ReflectionLoader();
+            pluginRegistrationObjectFactory = new PluginRegistrationObjectFactory();
+        }
+
+        public PluginRegistrationHelper(Action<string> logVerbose, Action<string> logWarning,
+            IReflectionLoader reflectionLoader, IPluginRegistrationObjectFactory pluginRegistrationObjectFactory)
+        {
+            this.logVerbose = logVerbose;
+            this.logWarning = logWarning;
+            this.reflectionLoader = reflectionLoader;
+            this.pluginRegistrationObjectFactory = pluginRegistrationObjectFactory;
+        }
+
+        public Assembly GetAssemblyRegistration(string assemblyName, string version) => pluginRepository.GetAssemblyRegistration(assemblyName, version);
 
         public void RemoveComponentsNotInMapping(Assembly assemblyMapping)
         {
-            var assemblyInCrm = pluginRepository.GetAssemblyRegistration(assemblyMapping.Name);
+            var assemblyInCrm = pluginRepository.GetAssemblyRegistration(assemblyMapping.Name, assemblyMapping.Version);
             if (assemblyInCrm == null)
             {
                 logVerbose?.Invoke($"Assembly {assemblyMapping.Name} not found in CRM");
@@ -62,12 +89,23 @@ namespace Xrm.Framework.CI.PowerShell.Cmdlets
             }
         }
 
-        public void DeleteObjectWithDependencies(Guid objectId, ComponentType? componentType)
+        public void DeleteObjectWithDependencies(Guid objectId, ComponentType? componentType, HashSet<string> deletingHashSet = null)
         {
-            logVerbose?.Invoke($"Checking dependecies for {componentType} / {objectId}");
+            if (deletingHashSet == null)
+            {
+                deletingHashSet = new HashSet<string>();
+            }
+            var objectkey = $"{componentType}{objectId}";
+            if (deletingHashSet.Contains(objectkey))
+            {
+                return;
+            }
+            deletingHashSet.Add(objectkey);
+
+            logVerbose?.Invoke($"Checking dependencies for {componentType} / {objectId}");
             foreach (var objectToDelete in GetDependeciesForDelete(objectId, componentType))
             {
-                DeleteObjectWithDependencies(objectToDelete.DependentComponentObjectId.Value, objectToDelete.DependentComponentTypeEnum);
+                DeleteObjectWithDependencies(objectToDelete.DependentComponentObjectId.Value, objectToDelete.DependentComponentTypeEnum, deletingHashSet);
             }
 
             switch (componentType)
@@ -84,15 +122,36 @@ namespace Xrm.Framework.CI.PowerShell.Cmdlets
                             Status = new OptionSetValue((int)Workflow_StatusCode.Draft)
                         });
                     }
+                    if (workflow.CategoryEnum == Workflow_Category.BusinessProcessFlow)
+                    {
+                        var entityMetadata = organizationService.GetEntityMetadata(workflow.UniqueName);
+                        logVerbose?.Invoke($"Checking dependencies for BPF entity: {workflow.UniqueName}");
+                        DeleteObjectWithDependencies(entityMetadata.MetadataId.Value, ComponentType.Entity, deletingHashSet);
+                    }
+
+                    if (workflow.CategoryEnum == Workflow_Category.BusinessProcessFlow)
+                    {
+                        RemoveAllWorkflowsFromBpf(workflow);
+                        logVerbose?.Invoke($"Preserving BPF {workflow.Name}");
+                        return;
+                    }
+
                     logVerbose?.Invoke($"Trying to delete {componentType} {workflow.Name}");
                     organizationService.Delete(Workflow.EntityLogicalName, objectId);
                     break;
                 case ComponentType.SDKMessageProcessingStep:
-                    logVerbose?.Invoke($"Trying to delete {componentType} {objectId}");
+                    var step = pluginRepository.GetSdkMessageProcessingStepById(objectId);
+                    if (step.IsHidden.Value == true)
+                    {
+                        logVerbose?.Invoke($"Preserving hidden SdkMessageProcessingStep {step.Name}");
+                        return;
+                    }
+                    logVerbose?.Invoke($"Trying to delete {componentType} {step.Name} / {objectId}");
                     organizationService.Delete(SdkMessageProcessingStep.EntityLogicalName, objectId);
                     break;
                 case ComponentType.PluginType:
-                    logVerbose?.Invoke($"Trying to delete {componentType} {objectId}");
+                    var type = pluginRepository.GetPluginTypeById(objectId);
+                    logVerbose?.Invoke($"Trying to delete {componentType} {type.Name} / {objectId}");
                     organizationService.Delete(PluginType.EntityLogicalName, objectId);
                     break;
                 case ComponentType.PluginAssembly:
@@ -104,6 +163,26 @@ namespace Xrm.Framework.CI.PowerShell.Cmdlets
                     organizationService.Delete(ServiceEndpoint.EntityLogicalName, objectId);
                     break;
             }
+        }
+
+        private const string ActionComposieClassWithAssemblyQualifiedName = "Microsoft.Crm.Workflow.Activities.ActionComposite, Microsoft.Crm.Workflow, Version=8.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35";
+        private const string mxswaNamespace = "clr-namespace:Microsoft.Xrm.Sdk.Workflow.Activities;assembly=Microsoft.Xrm.Sdk.Workflow, Version=8.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35";
+
+        private void RemoveAllWorkflowsFromBpf(Workflow bpf)
+        {
+            var xaml = XDocument.Parse(bpf.Xaml);
+            var nsmgr = new XmlNamespaceManager(new NameTable());
+            nsmgr.AddNamespace("mxswa", mxswaNamespace);
+            var actionsElements = xaml.XPathSelectElements($"//mxswa:ActivityReference[@AssemblyQualifiedName='{ActionComposieClassWithAssemblyQualifiedName}']", nsmgr).ToList();
+            foreach (var element in actionsElements)
+            {
+                element.Remove();
+            }
+            organizationService.Update(new Workflow
+            {
+                Xaml = xaml.ToString(SaveOptions.DisableFormatting),
+                Id = bpf.Id
+            });
         }
 
         public Guid UpsertPluginAssembly(Assembly pluginAssembly, AssemblyInfo assemblyInfo, string solutionName, RegistrationTypeEnum registrationType)
@@ -239,7 +318,7 @@ namespace Xrm.Framework.CI.PowerShell.Cmdlets
                 SASToken = serviceEndpt.SASToken,
                 UserClaimEnum = serviceEndpt.UserClaim,
                 Description = serviceEndpt.Description,
-                Url = serviceEndpt.Url,                
+                Url = serviceEndpt.Url,
                 AuthValue = serviceEndpt.AuthValue,
             };
 
@@ -276,7 +355,7 @@ namespace Xrm.Framework.CI.PowerShell.Cmdlets
                 FilteringAttributes = step.FilteringAttributes,
                 ImpersonatingUserId = new EntityReference(SystemUser.EntityLogicalName, pluginRepository.GetUserId(step.ImpersonatingUserFullname)),
                 ModeEnum = step.Mode,
-                SdkMessageFilterId = new EntityReference(SdkMessageFilter.EntityLogicalName, sdkMessageFilterId),
+                SdkMessageFilterId = sdkMessageFilterId.Equals(Guid.Empty) ? null : new EntityReference(SdkMessageFilter.EntityLogicalName, sdkMessageFilterId),
                 Rank = step.Rank,
                 StageEnum = step.Stage,
                 SupportedDeploymentEnum = step.SupportedDeployment,
@@ -369,11 +448,36 @@ namespace Xrm.Framework.CI.PowerShell.Cmdlets
             });
         }
 
-
         private IEnumerable<Dependency> GetDependeciesForDelete(Guid objectId, ComponentType? componentType) => ((RetrieveDependenciesForDeleteResponse)organizationService.Execute(new RetrieveDependenciesForDeleteRequest()
         {
             ComponentType = (int)componentType,
             ObjectId = objectId
         })).EntityCollection.Entities.Select(x => x.ToEntity<Dependency>());
+
+        public object GetPluginRegistrationObject(string assemblyPath, string customAttributeClass)
+        {
+            reflectionLoader.Initialise(assemblyPath, customAttributeClass);
+            return pluginRegistrationObjectFactory.GetAssembly(reflectionLoader);
+        }
+
+        public Assembly ReadMappingFile(string mappingFile)
+        {
+            var fileInfo = new FileInfo(mappingFile);
+            switch (fileInfo.Extension.ToLower())
+            {
+                case ".json":
+                    logVerbose("Reading mapping json file");
+                    var pluginAssembly = Serializers.ParseJson<Assembly>(mappingFile);
+                    logVerbose("Deserialized mapping json file");
+                    return pluginAssembly;
+                case ".xml":
+                    logVerbose("Reading mapping xml file");
+                    pluginAssembly = Serializers.ParseXml<Assembly>(mappingFile);
+                    logVerbose("Deserialized mapping xml file");
+                    return pluginAssembly;
+                default:
+                    throw new ArgumentException("Only .json and .xml mapping files are supported", nameof(ReadMappingFile));
+            }
+        }
     }
 }
